Refactor the options converting.
